+++ src/lv_widgets/lv_roller.h
@@ -150,7 +150,7 @@ uint16_t lv_roller_get_option_cnt(const lv_obj_t * roller);
  * @param buf pointer to an array to store the string
  * @param buf_size size of `buf` in bytes. 0: to ignore it.
  */
-void lv_roller_get_selected_str(const lv_obj_t * roller, char * buf, uint16_t buf_size);
+void lv_roller_get_selected_str(const lv_obj_t * roller, char * buf, uint32_t buf_size);
 
 /**
  * Get the align attribute. Default alignment after _create is LV_LABEL_ALIGN_CENTER
